// packages/database/convex/shared/generated_github_client.ts
export class ContentFileType extends S.Literal("file") {}

/**
 * Content File
 */
export class ContentFile extends S.Class<ContentFile>("ContentFile")({
	type: ContentFileType,
	encoding: S.String,
	size: S.Int,
	name: S.String,
	path: S.String,
	content: S.String,
	sha: S.String,
	url: S.String,
	git_url: S.NullOr(S.String),
	html_url: S.NullOr(S.String),
	download_url: S.NullOr(S.String),
	_links: S.Struct({
		git: S.NullOr(S.String),
		html: S.NullOr(S.String),
		self: S.String,
	}),
	target: S.optionalWith(S.String, { nullable: true }),
	submodule_git_url: S.optionalWith(S.String, { nullable: true }),
}) {}